dasca

+++ client/libs/movement/global.c
@@ -130,9 +130,9 @@ float get_ideal_angle(float target_angle) {
 }
 
 
-float get_current_time_ms() {
+double get_current_time_ms() {
     struct timespec spec;
 
     clock_gettime(CLOCK_MONOTONIC, &spec);
-    return ((float)spec.tv_nsec / 1.0e6);
+    return (spec.tv_nsec / 1.0e6);
 }
