+++ client/libs/movement/robotgrabballstep.c
@@ -5,14 +5,14 @@
 #include "../header/std_include.h"
 #include "../header/robotgrabballstep.h"
 
-static float begin_time;
+static double begin_time;
 
 #define GRABBER_SPEED 200
 #define GRABBER_TIME 700
 
 void robotgrabball_update(MotorInfo *motorInfo, SensorInfo *sensorInfo) {
     update_sensor_value(sensorInfo);
-    float current_time = get_current_time_ms();
+    double current_time = get_current_time_ms();
     printf("%f, %f", begin_time, current_time);
     if (global_params.robot_state == ROBOT_OPEN_GRABBER) {
         if (((current_time - begin_time)) > (GRABBER_TIME)) {
